updates to include more scenarios and C2 and poles
